requirements.txt
Flask==2.3.2
tensorflow==2.18.0
faiss-cpu==1.13.2
opencv-python==4.8.1.78
numpy==2.0.2
pillow==10.0.0
werkzeug==2.3.6
gunicorn==20.1.0